+++ resources/views/admin/news_categories/index.blade.php
@@ -1,3 +1,8 @@
+@if(!checkPer($_SESSION['user']['id'], 'post_view'))
+    <?php
+    header("Location: /superFood/admin/dashboard/");
+    ?>
+@endif
 @extends('admin.layouts.master')
 @section('title'){{'News Categories'}}@endsection
 @section('content')
@@ -11,8 +16,10 @@
                         <li class="breadcrumb-item"><a href="/superFood/admin/dashboard/">Dashboard</a></li>
                         <li class="breadcrumb-item active">Quản lý danh mục tin tức</li>
                     </ol>
+                    @if(checkPer($_SESSION['user']['id'], 'post_add'))
                     <a href="/superFood/admin/newsCategories/create" class="btn btn-primary addBtn">Thêm danh mục
                     </a>
+                    @endif
                     <div class="card mb-4">
                         <div class="card-header">
                             <i class="fas fa-table mr-1"></i>
@@ -34,8 +41,12 @@
                                         <td>{{$category->id}}</td>
                                         <td>{{$category->name}}</td>
                                         <td>
-                                            <a class="btn btn-primary" href="/superFood/admin/newsCategories/edit/{{$category->id}}">Sửa</a>
-                                            <a class="btn btn-danger" href="/superFood/admin/newsCategories/delete/{{$category->id}}">Xóa</a>
+                                            @if(checkPer($_SESSION['user']['id'], 'post_edit'))
+                                                <a class="btn btn-primary" href="/superFood/admin/newsCategories/edit/{{$category->id}}">Sửa</a>
+                                            @endif
+                                            @if(checkPer($_SESSION['user']['id'], 'post_delete'))
+                                                <a class="btn btn-danger" href="/superFood/admin/newsCategories/delete/{{$category->id}}">Xóa</a>
+                                            @endif
                                         </td>
                                     </tr>
                                     @endforeach
