+++ resources/views/admin/roles/index.blade.php
@@ -1,3 +1,8 @@
+@if(!checkPer($_SESSION['user']['id'], 'role_view'))
+    <?php
+    header("Location: /superFood/admin/dashboard/");
+    ?>
+@endif
 @extends('admin.layouts.master')
 @section('title'){{'News Categories'}}@endsection
 @section('content')
@@ -11,8 +16,10 @@
                         <li class="breadcrumb-item"><a href="/superFood/admin/dashboard/">Dashboard</a></li>
                         <li class="breadcrumb-item active">Quản lý phân quyền</li>
                     </ol>
-                    <a href="/superFood/admin/roles/create" class="btn btn-primary addBtn">Thêm quyền
-                    </a>
+                    @if(checkPer($_SESSION['user']['id'], 'role_add'))
+                        <a href="/superFood/admin/roles/create" class="btn btn-primary addBtn">Thêm quyền
+                        </a>
+                    @endif
                     <div class="card mb-4">
                         <div class="card-header">
                             <i class="fas fa-table mr-1"></i>
@@ -34,8 +41,12 @@
                                         <td>{{$role->id}}</td>
                                         <td>{{$role->name}}</td>
                                         <td>
+                                            @if(checkPer($_SESSION['user']['id'], 'role_edit'))
                                             <a class="btn btn-primary" href="/superFood/admin/roles/edit/{{$role->id}}">Sửa</a>
-                                            <a class="btn btn-danger" href="/superFood/admin/roles/delete/{{$role->id}}">Xóa</a>
+                                            @endif
+                                            @if(checkPer($_SESSION['user']['id'], 'role_delete'))
+                                                <a class="btn btn-danger" href="/superFood/admin/roles/delete/{{$role->id}}">Xóa</a>
+                                                @endif
                                         </td>
                                     </tr>
                                     @endforeach
