+++ storage/views/cache/e55e9cda7cacf82acfc0479f9a29c70df982af49.php
@@ -1,3 +1,8 @@
+<?php if(!checkPer($_SESSION['user']['id'], 'post_view')): ?>
+    <?php
+    header("Location: /superFood/admin/dashboard/");
+    ?>
+<?php endif; ?>
 
 <?php $__env->startSection('title'); ?><?php echo e('News Categories'); ?><?php $__env->stopSection(); ?>
 <?php $__env->startSection('content'); ?>
@@ -11,8 +16,10 @@
                         <li class="breadcrumb-item"><a href="/superFood/admin/dashboard/">Dashboard</a></li>
                         <li class="breadcrumb-item active">Quản lý danh mục tin tức</li>
                     </ol>
+                    <?php if(!checkPer($_SESSION['user']['id'], 'post_add')): ?>
                     <a href="/superFood/admin/newsCategories/create" class="btn btn-primary addBtn">Thêm danh mục
                     </a>
+                    <?php endif; ?>
                     <div class="card mb-4">
                         <div class="card-header">
                             <i class="fas fa-table mr-1"></i>
@@ -34,8 +41,12 @@
                                         <td><?php echo e($category->id); ?></td>
                                         <td><?php echo e($category->name); ?></td>
                                         <td>
-                                            <a class="btn btn-primary" href="/superFood/admin/newsCategories/edit/<?php echo e($category->id); ?>">Sửa</a>
-                                            <a class="btn btn-danger" href="/superFood/admin/newsCategories/delete/<?php echo e($category->id); ?>">Xóa</a>
+                                            <?php if(!checkPer($_SESSION['user']['id'], 'post_edit')): ?>
+                                                <a class="btn btn-primary" href="/superFood/admin/newsCategories/edit/<?php echo e($category->id); ?>">Sửa</a>
+                                            <?php endif; ?>
+                                            <?php if(!checkPer($_SESSION['user']['id'], 'post_delete')): ?>
+                                                <a class="btn btn-danger" href="/superFood/admin/newsCategories/delete/<?php echo e($category->id); ?>">Xóa</a>
+                                            <?php endif; ?>
                                         </td>
                                     </tr>
                                     <?php endforeach; $__env->popLoop(); $loop = $__env->getLastLoop(); ?>
