+++ storage/views/cache/e59d6f332f39daf0dde3618803a8e686a527bb3f.php
@@ -1,3 +1,8 @@
+<?php if(!checkPer($_SESSION['user']['id'], 'post_view')): ?>
+    <?php
+    header("Location: /superFood/admin/dashboard/");
+    ?>
+<?php endif; ?>
 
 <?php $__env->startSection('title'); ?><?php echo e('News Tags'); ?><?php $__env->stopSection(); ?>
 <?php $__env->startSection('content'); ?>
@@ -11,8 +16,10 @@
                         <li class="breadcrumb-item"><a href="/superFood/admin/dashboard/">Dashboard</a></li>
                         <li class="breadcrumb-item active">Quản lý Tags tin tức</li>
                     </ol>
+                    <?php if(!checkPer($_SESSION['user']['id'], 'post_add')): ?>
                     <a href="/superFood/admin/newsTags/create" class="btn btn-primary addBtn">Thêm Tag
                     </a>
+                    <?php endif; ?>
                     <div class="card mb-4">
                         <div class="card-header">
                             <i class="fas fa-table mr-1"></i>
@@ -34,8 +41,12 @@
                                         <td><?php echo e($tag->id); ?></td>
                                         <td><?php echo e($tag->name); ?></td>
                                         <td>
-                                            <a class="btn btn-primary" href="/superFood/admin/newsTags/edit/<?php echo e($tag->id); ?>">Sửa</a>
-                                            <a class="btn btn-danger" href="/superFood/admin/newsTags/delete/<?php echo e($tag->id); ?>">Xóa</a>
+                                            <?php if(!checkPer($_SESSION['user']['id'], 'post_edit')): ?>
+                                                <a class="btn btn-primary" href="/superFood/admin/newsTags/edit/<?php echo e($tag->id); ?>">Sửa</a>
+                                            <?php endif; ?>
+                                            <?php if(!checkPer($_SESSION['user']['id'], 'post_delete')): ?>
+                                                <a class="btn btn-danger" href="/superFood/admin/newsTags/delete/<?php echo e($tag->id); ?>">Xóa</a>
+                                            <?php endif; ?>
                                         </td>
                                     </tr>
                                     <?php endforeach; $__env->popLoop(); $loop = $__env->getLastLoop(); ?>
